+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## WIP
+ - Add 'Client Mods' settings icon
+
 ## r43
  - Make unique user reporting more robust, fixes discovery desync.